Separating out Matchers -- with changes in test file The files tests/test_main.py have been modified in accordance with the suricata/update/main.py and  suricata/update/matchers.py.
