+++ tests/test_main.py
@@ -117,7 +117,7 @@ class ModifyRuleFilterTestCase(unittest.TestCase):
     def test_id_match(self):
         rule0 = suricata.update.rule.parse(self.rule_string)
         line = '2020757 "\|0d 0a\|" "|ff ff|"'
-        rule_filter = main.ModifyRuleFilter.parse(line)
+        rule_filter = matchers.ModifyRuleFilter.parse(line)
         self.assertTrue(rule_filter != None)
         self.assertTrue(rule_filter.match(rule0))
         rule1 = rule_filter.filter(rule0)
@@ -128,7 +128,7 @@ class ModifyRuleFilterTestCase(unittest.TestCase):
     def test_re_match(self):
         rule0 = suricata.update.rule.parse(self.rule_string)
         line = 're:classtype:trojan-activity "\|0d 0a\|" "|ff ff|"'
-        rule_filter = main.ModifyRuleFilter.parse(line)
+        rule_filter = matchers.ModifyRuleFilter.parse(line)
         self.assertTrue(rule_filter != None)
         self.assertTrue(rule_filter.match(rule0))
         rule1 = rule_filter.filter(rule0)
@@ -139,7 +139,7 @@ class ModifyRuleFilterTestCase(unittest.TestCase):
     def test_re_backref_one(self):
         rule0 = suricata.update.rule.parse(self.rule_string)
         line = 're:classtype:trojan-activity "(alert)(.*)" "drop\\2"'
-        filter = main.ModifyRuleFilter.parse(line)
+        filter = matchers.ModifyRuleFilter.parse(line)
         self.assertTrue(filter != None)
         self.assertTrue(filter.match(rule0))
         rule1 = filter.filter(rule0)
@@ -149,7 +149,7 @@ class ModifyRuleFilterTestCase(unittest.TestCase):
     def test_re_backref_two(self):
         rule0 = suricata.update.rule.parse(self.rule_string)
         line = 're:classtype:trojan-activity "(alert)(.*)(from_server)(.*)" "drop\\2to_client\\4"'
-        filter = main.ModifyRuleFilter.parse(line)
+        filter = matchers.ModifyRuleFilter.parse(line)
         self.assertTrue(filter != None)
         self.assertTrue(filter.match(rule0))
         rule1 = filter.filter(rule0)
@@ -160,7 +160,7 @@ class ModifyRuleFilterTestCase(unittest.TestCase):
         rule_in = suricata.update.rule.parse(self.rule_string)
         self.assertIsNotNone(rule_in)
 
-        f = main.ModifyRuleFilter.parse(
+        f = matchers.ModifyRuleFilter.parse(
             'group:emerging-trojan.rules "^alert" "drop"')
         self.assertIsNotNone(f)
 
@@ -168,14 +168,14 @@ class ModifyRuleFilterTestCase(unittest.TestCase):
         self.assertTrue(rule_out.format().startswith("drop"))
 
     def test_oinkmaster_backticks(self):
-        f = main.ModifyRuleFilter.parse(
+        f = matchers.ModifyRuleFilter.parse(
             '* "^drop(.*)noalert(.*)" "alert${1}noalert${2}"')
         rule_in ="""drop http $EXTERNAL_NET any -> $HOME_NET any (msg:"ET MALWARE Windows executable sent when remote host claims to send an image 2"; flow: established,to_client; content:"|0d 0a|Content-Type|3a| image/jpeg|0d 0a 0d 0a|MZ"; fast_pattern:12,20; noalert; classtype:trojan-activity; sid:2020757; rev:2;)"""
         rule_out = f.filter(suricata.update.rule.parse(rule_in))
         self.assertEqual("""alert http $EXTERNAL_NET any -> $HOME_NET any (msg:"ET MALWARE Windows executable sent when remote host claims to send an image 2"; flow: established,to_client; content:"|0d 0a|Content-Type|3a| image/jpeg|0d 0a 0d 0a|MZ"; fast_pattern:12,20; noalert; classtype:trojan-activity; sid:2020757; rev:2;)""", rule_out.format())
 
     def test_oinkmaster_backticks_not_noalert(self):
-        f = main.ModifyRuleFilter.parse(
+        f = matchers.ModifyRuleFilter.parse(
             'modifysid * "^drop(.*)noalert(.*)" | "alert${1}noalert${2}"')
         rule_in ="""drop http $EXTERNAL_NET any -> $HOME_NET any (msg:"ET MALWARE Windows executable sent when remote host claims to send an image 2"; flow: established,to_client; content:"|0d 0a|Content-Type|3a| image/jpeg|0d 0a 0d 0a|MZ"; fast_pattern:12,20; classtype:trojan-activity; sid:2020757; rev:2;)"""
         rule_out = f.filter(suricata.update.rule.parse(rule_in))
@@ -183,7 +183,7 @@ class ModifyRuleFilterTestCase(unittest.TestCase):
 
     def test_oinkmaster_modify_group_name(self):
         """Test an Oinkmaster style modification line using a group name."""
-        f = main.ModifyRuleFilter.parse(
+        f = matchers.ModifyRuleFilter.parse(
             'modifysid botcc.rules "^alert" | "drop"')
         rule_in ="""alert http $EXTERNAL_NET any -> $HOME_NET any (msg:"ET MALWARE Windows executable sent when remote host claims to send an image 2"; flow: established,to_client; content:"|0d 0a|Content-Type|3a| image/jpeg|0d 0a 0d 0a|MZ"; fast_pattern:12,20; classtype:trojan-activity; sid:2020757; rev:2;)"""
         rule = suricata.update.rule.parse(rule_in, "rules/botcc.rules")
